项目展示完成

+++ crowdfunding07-member-parent/crowdfunding09-member-entity/src/main/java/com/example/crowdfunding/bean/vo/ProjectVO.java
@@ -32,7 +32,7 @@ public class ProjectVO implements Serializable {
     private String projectDescription;
 
     // 计划筹集的金额
-    private Integer money;
+    private long money;
 
     // 筹集资金的天数
     private Integer day;
